+++ package.json
@@ -11,7 +11,6 @@
     "@types/react-transition-group": "^4.4.4",
     "cloudinary-react": "^1.7.1",
     "formik": "^2.2.9",
-    "history": "^5.3.0",
     "js-cookie": "^3.0.1",
     "jwt-decode": "^3.1.2",
     "material-ui-image": "^3.3.2",
