+++ src/common/testUtils.tsx
@@ -1,25 +1,13 @@
-import { RenderOptions, render } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import { ThemeProvider } from "@mui/material/styles";
-import { EnhancedStore, configureStore } from "@reduxjs/toolkit";
+import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { ReactElement, ReactNode } from "react";
 import { Form, Formik, FormikConfig } from "formik";
-import { RootState, reducer } from "store";
+import { reducer } from "store";
 import theme from "theme";
-
-interface RenderProps extends RenderOptions {
-  readonly preloadedState: Partial<RootState>;
-  readonly store: EnhancedStore;
-}
-
-interface WrapperProps {
-  readonly children: ReactElement;
-}
-
-interface FormValues {
-  readonly [key: string]: string;
-}
+import { RenderProps, StringMap, WrapperProps } from "./types";
 
 /**
  * Wraps the element being tested with Theme, Provider, and
@@ -58,7 +46,7 @@ export const renderWithContext = (
 
 export const withFormik = (
   element: ReactNode,
-  props: FormikConfig<FormValues>
+  props: FormikConfig<StringMap>
 ) => {
   return (
     <Formik { ...props }>
